moduleAI/requirements.txt
transformers
torch
requests
python-docx
python-dotenv
pydantic
